Avoid timestamp overflow in recommendations
