+++ src/main/java/com/exampleepam/restaurant/service/forecast/MonthlyResult.java
@@ -1,0 +1,13 @@
+package com.exampleepam.restaurant.service.forecast;
+
+import java.time.YearMonth;
+import java.util.List;
+import java.util.Map;
+
+import com.exampleepam.restaurant.service.forecast.ForecastResult;
+
+public record MonthlyResult(ScaleData scale,
+                            Map<YearMonth, Integer> monthForecasts,
+                            List<Integer> modelHistory,
+                            ForecastResult result,
+                            boolean singlePoint) {}
