+++ src/main/java/com/exampleepam/restaurant/service/forecast/ScaleData.java
@@ -1,0 +1,5 @@
+package com.exampleepam.restaurant.service.forecast;
+
+import java.util.List;
+
+public record ScaleData(List<String> labels, List<Integer> actual, List<Integer> forecast) {}
