+++ src/main/resources/templates/dish-forecast-info.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8"/>
+    <title>Forecasting Help</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" crossorigin="anonymous">
+</head>
+<body class="bg-dark text-light">
+<div class="container my-4">
+    <h2>Forecasting models</h2>
+    <p>This page explains the forecasting options for dishes.</p>
+    <ul>
+        <li><strong>Holt&#8209;Winters</strong>: triple exponential smoothing capturing level, trend and seasonality. Works well for steady demand with gradual changes.</li>
+        <li><strong>ARIMA(1,0,0)</strong>: a simple autoregressive model using the previous value to project the next. Suitable for short histories without seasonality.</li>
+        <li><strong>Auto&#8209;ARIMA</strong>: searches a range of ARIMA parameters and selects the best according to AIC.</li>
+    </ul>
+    <p>When only a single non‑zero month is available, the system repeats that value for all future months.</p>
+    <a href="/admin/dish-forecast" class="btn btn-outline-light mt-3">Back to forecasts</a>
+</div>
+</body>
+</html>
